Separated updateSvg
Only works for walkers

Modularized js files by separating the frontend and backend of the svg drawing

Also changed the target offset to be more accurate to the middle of the circle

Fullscreening does not work; breaks how drawing works.
