+++ update_svg.js
@@ -1,0 +1,155 @@
+var parentSvg;
+var textSvg;
+var svgElements;
+
+
+function startSvg() {
+    parentSvg = document.getElementById("parent_svg");
+    textSvg = document.getElementById("debugging");
+    svgElements = {};
+}
+
+
+function updateSvg(worldModel, screen_width, screen_height){
+    let xOffset = 0;
+    let yOffset = 0;
+    let width = 0;
+    let height = 0;  
+    let tooltipInfo;
+
+    var gElement = document.getElementById("tooltip-group");
+    if(!gElement){
+        gElement = document.createElementNS("http://www.w3.org/2000/svg", "g");
+    
+        // Set the attributes of the g element
+        gElement.setAttribute("id", "tooltip-group");
+        gElement.setAttribute("visibility", "hidden");
+
+        var textElement = document.createElementNS("http://www.w3.org/2000/svg", "text");
+
+        // Set the attributes of the text element
+        textElement.setAttribute("id", "tooltip");
+        textElement.setAttribute("width", "50");
+        textElement.setAttribute("height", "50");
+        textElement.setAttribute("fill", "black");
+        textElement.setAttribute("background-color", "#000000");
+         // Create a <rect> element to serve as the background of the tooltip
+         var rectElement = document.createElementNS("http://www.w3.org/2000/svg", "rect");
+         rectElement.setAttribute("fill", "lightgray");
+         rectElement.setAttribute("rx", 5);
+         rectElement.setAttribute("ry", 5);
+         rectElement.setAttribute("id", "tooltipBox");
+
+        gElement.appendChild(rectElement);
+        gElement.appendChild(textElement);
+        parentSvg.appendChild(gElement);
+    }
+    Object.entries(worldModel["key"]).forEach(([key, value]) => {
+      if (key != null && value["screenRect"] != null) {
+        xOffset = value["screenRect"].x/screen_width*100;
+        yOffset = value["screenRect"].y/screen_height*100;
+        width = value["screenRect"].w/screen_width*100;
+        height = value["screenRect"].h/screen_height*100;
+
+        var svgRect = svgElements[key];
+        if (svgRect) {
+            svgRect.setAttribute("width", width.toString()+"%");
+            svgRect.setAttribute("height", height.toString()+"%");
+            svgRect.setAttribute("x", xOffset.toString()+"%");
+            svgRect.setAttribute("y", yOffset.toString()+"%");
+        }
+        else{
+            svgRect = document.createElementNS(
+                "http://www.w3.org/2000/svg",
+                "rect"
+            );
+            svgRect.setAttribute("id", key);
+            svgRect.setAttribute("width", width.toString()+"%");
+            svgRect.setAttribute("height", height.toString()+"%");
+            svgRect.setAttribute("x", xOffset.toString()+"%");
+            svgRect.setAttribute("y", yOffset.toString()+"%");
+            svgRect.setAttribute("fill", "none");
+            svgRect.setAttribute("stroke", "red");
+            svgRect.setAttribute("stroke-width", "2");
+            svgRect.setAttribute("position", "absolute");
+            svgRect.style.pointerEvents = "all"; // prevent stroke from triggering mouse events    
+            svgRect.addEventListener("mousemove", (evt) => {
+                // Set the text content of the text element
+                var CTM = svgRect.getScreenCTM();
+                var mouseX = (evt.clientX - CTM.e) / CTM.a;
+                var mouseY = (evt.clientY - CTM.f) / CTM.d;
+            
+                tooltipInfo = Object.entries(value).map(function(entry) {
+                    var key = entry[0];
+                    var value = entry[1];
+                    if (typeof value === "object") {
+                        return key + ": " + JSON.stringify(value);
+                    } else {
+                        return key + ": " + value;
+                    }
+                }).join("\n");
+            
+                // Split the tooltip info into an array of lines
+                var tooltipLines = tooltipInfo.split("\n");
+            
+                gElement.setAttribute("transform", `translate(${mouseX + 6 / CTM.a}, ${mouseY + 20 / CTM.d})`);
+                gElement.setAttribute("class", "tooltip");
+                if(!textElement){
+                    textElement =  document.getElementById("tooltip");
+                }
+                textElement.textContent = "";
+        
+                tooltipLines.forEach(function(line, index) {
+                    var tspan = document.createElementNS("http://www.w3.org/2000/svg", "tspan");
+                    tspan.setAttribute("x", 0);
+                    tspan.setAttribute("dy", "1.2em"); // Line spacing
+                    tspan.textContent = line;
+                    textElement.appendChild(tspan);
+                });
+
+                // Set the attributes of the rect element
+                var bbox = textElement.getBBox();
+                if(!rectElement){
+                    rectElement = document.getElementById("tooltipBox");
+                }
+                rectElement.setAttribute("width", bbox.width + 10);
+                rectElement.setAttribute("height", bbox.height + 10);
+                rectElement.setAttribute("x", bbox.x - 5);
+                rectElement.setAttribute("y", bbox.y - 5);
+
+                gElement.setAttribute("visibility", "visible");
+                // console.log("mouse in");
+            });            
+            svgRect.addEventListener("mouseout", () => {
+                gElement.setAttribute("visibility", "hidden");
+                console.log("mouse out");
+            });
+            svgRect.addEventListener("click", (evt) => {
+                evt.target.setAttribute("stroke", getRandomColor());
+            });
+        
+            parentSvg.appendChild(svgRect);
+            svgElements[key] = svgRect;
+        }
+      }
+    });
+
+    //First thing to do: if an item is in the svgElements list and is no longer in the worldModel. remove it
+    Object.entries(svgElements).forEach(([key, value]) => {
+        if (!(key in worldModel["key"])) {
+            if (parentSvg.contains(value)) {
+                parentSvg.removeChild(value);
+            }
+            delete svgElements[key];
+        }
+    });
+    
+
+}
+
+function getRandomColor() {
+    const colors = ['#e6194b', '#3cb44b', '#ffe119', '#0082c8', '#f58231', '#911eb4'];
+    return colors[Math.floor(Math.random() * colors.length)];
+}
+
+export { startSvg, updateSvg };
