+++ video_overlay.html
@@ -6,7 +6,8 @@
   <script src="https://extension-files.twitch.tv/helper/v1/twitch-ext.min.js"></script>
   <!-- <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js" type="text/javascript"></script> -->
   <script src="jquery.min.js"></script>
-  <script src="video_overlay.js" type="text/javascript"></script>
+  <script type="module" src="video_overlay.js" type="text/javascript"></script>
+  <script src="update_svg.js" type="text/javascript"></script>
 </head>
 <body style="font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif; width: 100%; height: 100%;">
     <div id="app" class="full-height"></div>
